implements CustomerService interface and add contextLoads test
